requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
django-rest-framework-simplejwt==5.4.0
gunicorn==23.0.0
packaging==25.0
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
